working password type
